feat: function for filtering added - filterUtils.js
