+++ src/routes/agentRoutesAPI.js
@@ -28,7 +28,6 @@ import { getProfitLoss } from "../controller/agentProfitLossController.js";
 import { getUserById } from "../controller/userController/getUserById.js";
 import { getClientSummary } from "../controller/agentClientSummaryController.js";
 import { generateUserIdCommissionLimit } from "../controller/generateUserIdCommissionLimit.js";
-import { fetchTableData } from "../utils/fetchTableData.js";
 
 const router = express.Router();
 
@@ -76,7 +75,5 @@ router.get("/user/:id", getUserById);
 router.get("/blocked", getBlockedClients); //http://localhost:4320/auth-api/agent/blocked
 
 router.post("/generateUserIdCommissionLimit", generateUserIdCommissionLimit); //http://localhost:4320/auth-api/agent/generateUserIdCommissionLimit
-
-router.get("/filteredData", fetchTableData); //http://localhost:4320/auth-api/agent/filteredData?table=agents&userId=6&&limit=10&offset=0&sortBy=userId&sortOrder=DESC
 
 export default router;
